Move tests to main test_geocode module
Previously tests were in tools/tests in
separate to the main tests/test_geocode module
